+++ packages/gatsby-design-tokens/src/borders.js
@@ -1,0 +1,1 @@
+export default [0, `1px solid`, `2px solid`]
