--- www/src/utils/tokens/breakpoints.js
@@ -1,9 +1,0 @@
-export default {
-  xxs: 0, // needed for styled-system
-  xs: `400px`,
-  sm: `550px`,
-  md: `750px`,
-  lg: `1000px`,
-  xl: `1200px`,
-  xxl: `1600px`,
-}
