--- www/src/utils/tokens/font-weights.js
@@ -1,1 +1,0 @@
-export default [400, 700, 800]
