--- www/src/utils/tokens/index.js
@@ -1,31 +1,0 @@
-import breakpoints from "./breakpoints"
-import mediaQueries from "./media-queries"
-import colors from "./colors"
-import sizes from "./sizes"
-import letterSpacings from "./letter-spacings"
-import lineHeights from "./line-heights"
-import radii from "./radii"
-import shadows from "./shadows"
-import space from "./space"
-import transition from "./transition"
-import fontSizes from "./font-sizes"
-import fonts from "./fonts"
-import zIndices from "./z-indices"
-import fontWeights from "./font-weights"
-
-export {
-  breakpoints,
-  mediaQueries,
-  colors,
-  sizes,
-  letterSpacings,
-  lineHeights,
-  radii,
-  shadows,
-  space,
-  transition,
-  fontSizes,
-  fonts,
-  fontWeights,
-  zIndices,
-}
