--- www/src/utils/tokens/letter-spacings.js
@@ -1,5 +1,0 @@
-export default {
-  normal: `normal`,
-  tracked: `0.075em`,
-  tight: `-0.015em`,
-}
