--- www/src/utils/tokens/line-heights.js
@@ -1,6 +1,0 @@
-export default {
-  solid: 1,
-  dense: 1.25,
-  default: 1.5,
-  loose: 1.75,
-}
