--- www/src/utils/tokens/media-queries.js
@@ -1,8 +1,0 @@
-import breakpoints from "./breakpoints"
-
-let mediaQueries = {}
-for (let breakpoint in breakpoints) {
-  mediaQueries[breakpoint] = `@media (min-width: ${breakpoints[breakpoint]})`
-}
-
-export default mediaQueries
