--- www/src/utils/tokens/radii.js
@@ -1,1 +1,0 @@
-export default [0, 2, 4, 8, 16, 9999, `100%`]
