--- www/src/utils/tokens/sizes.js
@@ -1,11 +1,0 @@
-export default {
-  headerHeight: `3.75rem`,
-  bannerHeight: `2.5rem`,
-  sidebarUtilityHeight: `3.5rem`,
-  pageHeadingDesktopWidth: `3.5rem`,
-  sidebarWidth: {
-    default: 11,
-    large: 12,
-  },
-  avatar: `2rem`,
-}
