--- www/src/utils/tokens/space.js
@@ -1,1 +1,0 @@
-export default [0, 4, 8, 12, 16, 20, 24, 32, 40, 48, 56, 64, 72]
