--- www/src/utils/tokens/transition.js
@@ -1,10 +1,0 @@
-export default {
-  curve: {
-    default: `cubic-bezier(0.4, 0, 0.2, 1)`,
-  },
-  speed: {
-    default: `250ms`,
-    fast: `100ms`,
-    slow: `350ms`,
-  },
-}
