--- www/src/utils/tokens/z-indices.js
@@ -1,9 +1,0 @@
-export default {
-  feedbackWidget: 2,
-  navigation: 5,
-  banner: 10,
-  modal: 10,
-  sidebar: 10,
-  sidebarToggleButton: 20,
-  skipLink: 100,
-}
